% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_pkgs.R
\name{list_pkgs}
\alias{list_pkgs}
\title{List packages and their paths}
\usage{
list_pkgs()
}
\value{
NULL if no packages are loaded.
}
\description{
This function lists the packages and their paths that have been loaded
using the \code{multiloadr} option. For each package, it also displays the
name of the current branch in the Git repository located at the package path.
}
\examples{
\dontrun{
list_pkgs()
}

}
