% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pkgs.R
\name{load_pkgs}
\alias{load_pkgs}
\title{Load packages listed in the multiloadr option}
\usage{
load_pkgs()
}
\description{
This function loads packages listed in the multiloadr option using the \code{devtools::load_all()} function.
}
\examples{
load_pkgs()

}
