% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiloadr-package.R
\docType{package}
\name{multiloadr-package}
\alias{multiloadr}
\alias{multiloadr-package}
\title{multiloadr: Load the packages relevant to a project within a session.}
\description{
The purpose of this package is to assist in loading the R packages that serve as immediate dependencies for the R package currently being worked on. This package is useful for developing multiple interdependent R packages.
}
\keyword{internal}
