% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_pkgs.R
\name{rm_pkgs}
\alias{rm_pkgs}
\title{Remove packages from \code{multiloadr} configuration}
\usage{
rm_pkgs(pkg_names)
}
\arguments{
\item{pkg_names}{A character vector specifying the names of the packages
to remove.}
}
\description{
This function removes specified package(s) from the \code{multiloadr}
configuration.
}
\examples{
rm_pkgs(c("package1", "package2"))

}
