% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pkgs.R
\name{add_pkgs}
\alias{add_pkgs}
\title{Add a package path to the multiloadr options}
\usage{
add_pkgs(pkg_name, path)
}
\arguments{
\item{pkg_name}{The name of the package to add.}

\item{path}{The path to the package.}
}
\value{
This function is called for its side effects.
}
\description{
This function adds a package path to the \code{multiloadr} option, which allows R
to load multiple packages in a single session. The \code{multiloadr} option is a
named list of package paths, where the name of each entry corresponds to the
package name.
}
\examples{
# Add package path to multiloadr option
add_pkgs("dplyr", "/path/to/dplyr")

}
