% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pkgs.R
\name{load_pkgs}
\alias{load_pkgs}
\title{Load packages and switch git branch if specified}
\usage{
load_pkgs(branch_name = NULL)
}
\arguments{
\item{branch_name}{Character string specifying the name of the git branch to
switch to. Defaults to \code{NULL}. In case the branch is NULL or
non-existent, it will be loaded from the presently active branch.}
}
\value{
This function returns nothing, but prints a message indicating which
packages were loaded and from which branch.
}
\description{
This function loads packages specified in the \code{multiloadr} option and
switches to a specified git branch if provided.
}
\examples{
# Load packages without switching branch
load_pkgs()

# Load packages from "develop" branch
load_pkgs("develop")
}
