% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiloadr-package.R
\docType{package}
\name{multiloadr-package}
\alias{multiloadr}
\alias{multiloadr-package}
\title{\code{multiloadr}: seamlessly load multiple R packages}
\description{
A versatile workflow tool that simplifies the loading of multiple
R packages seamlessly, offering enhanced functionality across different
branches
}
\author{
\strong{Maintainer}: Dony Unardi \email{donyunardi@gmail.com}

}
\keyword{internal}
